Edited CLI mainly.
